Move ui elements to ui folder
