+++ components.json
@@ -13,7 +13,7 @@
   "aliases": {
     "components": "@/components",
     "utils": "@/lib/utils",
-    "ui": "@/components/ui_elements",
+    "ui": "@/components/ui/ui_elements",
     "lib": "@/lib",
     "hooks": "@/hooks"
   }
