+++ src/components/TaskLabel/DatePicker.tsx
@@ -1,9 +1,9 @@
 import { Dispatch, SetStateAction, useEffect, useState } from "react";
-import { Popover, PopoverAnchor, PopoverContent, PopoverTrigger } from "../ui_elements/popover";
+import { Popover, PopoverAnchor, PopoverContent, PopoverTrigger } from "../ui/ui_elements/popover";
 import { format, isValid, parse } from "date-fns";
-import { Input } from "../ui_elements/input";
+import { Input } from "../ui/ui_elements/input";
 import { CalendarIcon } from "lucide-react";
-import { Calendar } from "../ui_elements/calendar";
+import { Calendar } from "../ui/ui_elements/calendar";
 
 export interface DateInputProps {
   date: Date | undefined,
